Skip hanging spawn tests (#10838)
